fix: prevent program from extending cancelled HITs
If we are using FixedTimeoutPolicy, the newly created batch will have the same hit_key. However, the HITState is cancelled so extending it will lead to failure. We should instead post a new HIT.

In addition, GroupID is now changed to UUID of each question (instead of title) so they are different during each run.