fix (SurveyPolicy): Fix Jonker-Volgenant infinite loop error && ...
- hacks around the J-V algorithm infinite loop problem
- fixes a bug where arrays are initialized with incorrect length.
- includes a separate Kuhn–Munkres algorithm (Hungarian algorithm) as backup

The Jonker-Volgenant algorithm sometimes ends up in an infinite loop.
This error also happens to the original CPP code and is acknowledged by Volgenant himself [1].
[1] suggests to scale up the cost matrix by a constant factor. However, it does not work in our case, as some distances are very close to zero and remain so after scaling up (e.g. 1E-7)

This (slightly ugly) fix is inspired by an implementation in JavaScript [2].

As a side note, it seems that the current popular way of solving assignment problem is using scipy [3].
It does some major modifications over J-V algorithm based on a 2016 paper [4].
This could potentially be a future to-do item.

References
[1] https://github.com/hrldcpr/pyLAPJV#nota-bene
[2] https://github.com/Fil/lap-jv/blob/1e5857b5e7d764ced8f760eaac1657d5cc234819/lap.js
[3] https://github.com/scipy/scipy/blob/v1.8.1/scipy/optimize/rectangular_lsap/rectangular_lsap.cpp
[4] Crouse. On implementing 2D rectangular assignment algorithms. https://doi.org/10.1109/TAES.2016.140952